Still hammering out backend.
Strange issue with backend only compiling properly the first time after any change is made to the backend files.

Server-side rendering and routing system is getting there.  Issues now reside in src/layout file's logic for the navigation bar.  E.g. window variable is called in a function in Admin.jsx to get the current URL, but window is not defined from SSR, so another way will need to be found.